Version 1.1.6

+++ src/C/amd.c
@@ -1,9 +1,9 @@
 /*
- * Copyright 2012 M. Andersen and L. Vandenberghe.
+ * Copyright 2012-2013 M. Andersen and L. Vandenberghe.
  * Copyright 2010-2011 L. Vandenberghe.
  * Copyright 2004-2009 J. Dahl and L. Vandenberghe.
  *
- * This file is part of CVXOPT version 1.1.5.
+ * This file is part of CVXOPT version 1.1.6.
  *
  * CVXOPT is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -156,7 +156,7 @@ static PyObject* order_c(PyObject *self, PyObject *args, PyObject *kwrds)
         return NULL;
     }
     if (uplo != 'U' && uplo != 'L') err_char("uplo", "'L', 'U'");
-    if (!(perm = (matrix *) Matrix_New(SP_NROWS(A),1,INT)))
+    if (!(perm = (matrix *) Matrix_New((int)SP_NROWS(A),1,INT)))
         return PyErr_NoMemory();
     n = SP_NROWS(A);
     for (nnz=0, j=0; j<n; j++) {
